+++ ShadowApiNet/HttpHandlers/HttpPutHandler.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Newtonsoft.Json;
-using ShadowApiNet.Interfaces;
+using ShadowApiNet.Abstractions;
 using System;
 using System.Collections.Generic;
 using System.IO;
